permissions OK for book update & book delete

+++ Booker/src/main/java/ch/hearc/qdljee/controllers/BooksController.java
@@ -5,6 +5,7 @@ import java.io.FileOutputStream;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.sql.Date;
+import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -32,6 +33,7 @@ import ch.hearc.qdljee.dto.BookDto;
 import ch.hearc.qdljee.dto.CommentDto;
 import ch.hearc.qdljee.dto.SearchDto;
 import ch.hearc.qdljee.model.Books;
+import ch.hearc.qdljee.model.Role;
 import ch.hearc.qdljee.model.User;
 import ch.hearc.qdljee.service.BookService;
 import ch.hearc.qdljee.service.CommentService;
@@ -142,8 +144,11 @@ public class BooksController {
 
 	@PostMapping("/{id}/delete")
 	public String deleteBook(Model model, @PathVariable("id") int id) {
+		if(!validatePermForBook(id)) {
+			return "redirect:/Books/?errorPermissions";
+		}
 		String path = environment.getProperty("images.path");
-		bookService.delete(id,path);
+		bookService.delete(id, path);
 		return "redirect:/Books";
 	}
 
@@ -156,6 +161,10 @@ public class BooksController {
 
 	@PostMapping("/{id}/update")
 	public String updateBook(Model model, @PathVariable("id") int id, @ModelAttribute("Bookdto") BookDto bookDto) {
+		if(!validatePermForBook(id)) {
+			return "redirect:/Books/?errorPermissions";
+		}
+
 		String imageURL = null;
 		if (bookDto.getImage() != null && bookDto.getImage().getOriginalFilename().contains("octet-stream")) {
 			String path = environment.getProperty("images.path");
@@ -183,6 +192,21 @@ public class BooksController {
 			e.printStackTrace();
 		}
 		return "redirect:/Books/" + id;
+	}
+
+	private boolean validatePermForBook(int id) {
+		Collection<Role> roles = Tools.getCurrentUser().getRoles();
+		boolean status = false;
+		for (Role role : roles) {
+			if (!role.getName().equals("ROLE_AUTHOR")
+					&& !bookService.getBooksById(id).getAuthor().equals(Tools.getCurrentUser())) {
+				status = true;
+			}
+			if (role.getName().equals("ROLE_ADMIN")) {
+				status = false;
+			}
+		}
+		return !status;
 	}
 
 	@PostMapping("/{id}/createComment")
